Implemented saving and retrieving students from database

+++ app/src/main/java/me/dbecaj/friurnik/data/interactors/schedule/ScheduleInteractor.java
@@ -13,6 +13,6 @@ public interface ScheduleInteractor {
         void failure(String error);
     }
 
-    void getSchedule(int studentId, ScheduleListener listener);
+    void getSchedule(long studentId, ScheduleListener listener);
 
 }
