+++ app/src/main/java/me/dbecaj/friurnik/ui/schedule/ScheduleMvp.java
@@ -14,7 +14,7 @@ public interface ScheduleMvp {
 
     interface Presenter extends BaseMvp.Presenter {
 
-        void loadSchedule(int studentId);
+        void loadSchedule(long studentId);
     }
 
 }
